+++ mailjet/client.py
@@ -7,6 +7,7 @@ import logging
 import requests
 from requests.compat import urljoin
 
+requests.packages.urllib3.disable_warnings()
 
 class Config(object):
     API_URL = 'https://api.mailjet.com/v3/REST/'
@@ -92,31 +93,31 @@ class Config(object):
 
 class Endpoint(object):
 
-    def __init__(self, url, doc, auth):
-        self._url, self._doc, self._auth = url, doc, auth
+    def __init__(self, url, doc, auth, action=None):
+        self._url, self._doc, self._auth, self.action = url, doc, auth, action
 
     def __doc__(self):
         return self._doc
 
-    def _get(self, res_id=None, **kwargs):
-        return api_call(self._auth, 'get', self._url, resource_id=res_id, **kwargs)
+    def _get(self, filters=None, action_id=None, id=None, **kwargs):
+        return api_call(self._auth, 'get', self._url, action=self.action, action_id=action_id, filters=filters, resource_id=id, **kwargs)
 
-    def get_many(self, **kwargs):
-        return self._get(**kwargs)
+    def get_many(self, filters=None, action_id=None, **kwargs):
+        return self._get(filters=filters, **kwargs)
 
-    def get(self, res_id, **kwargs):
-        return self._get(res_id, **kwargs)
+    def get(self, id=None, filters=None, action_id=None, **kwargs):
+        return self._get(id=id, filters=filters, **kwargs)
 
-    def create(self, data, **kwargs):
-        return api_call(self._auth, 'post', self._url, data=data, **kwargs)
+    def create(self, data, filters=None, action_id=None, **kwargs):
+        return api_call(self._auth, 'post', self._url, data=data, action=self.action, action_id=action_id, filters=filters, **kwargs)
 
     new = create
 
-    def update(self, res_id, data, **kwargs):
-        return api_call(self._auth, 'put', self._url, resource_id=res_id, data=data, **kwargs)
+    def update(self, res_id, data, filters=None, action_id=None, **kwargs):
+        return api_call(self._auth, 'put', self._url, resource_id=res_id, data=data, action=self.action, action_id=action_id, filters=filters, **kwargs)
 
     def delete(self, res_id, **kwargs):
-        return api_call(self._auth, 'delete', self._url, resource_id=res_id, **kwargs)
+        return api_call(self._auth, 'delete', self._url, action=self.action, action_id=action_id, resource_id=res_id, **kwargs)
 
 
 class Client(object):
@@ -129,21 +130,25 @@ class Client(object):
         return self.endpoints
 
     def __getattr__(self, name):
+        split = name.split('_')
+        name = split[0]
+        if (len(split) > 1):
+            action = split[1]
         if name in self.endpoints:
             url, doc = self.config[name]
-            return type(name, (Endpoint,), {})(url=url, doc=doc, auth=self.auth)
+            return type(name, (Endpoint,), {})(url=url, doc=doc, action=action, auth=self.auth)
         raise AttributeError
 
 
-def api_call(auth, method, url, data=None, resource_id=None, extra_headers=None,
-             timeout=60, debug=False, **kwargs):
-    data, url = build_body_and_url(url, method=method,
-                                   data=data, resource_id=resource_id)
+def api_call(auth, method, url, data=None, filters=None, resource_id=None, extra_headers=None,
+             timeout=60, debug=False, action=None, action_id=None, **kwargs):
+    url = build_url(url, method=method, action=action, resource_id=resource_id, action_id=action_id)
+    print method, url, data
     headers = build_headers(extra_headers)
     req_method = getattr(requests, method)
 
     try:
-        response = req_method(url, data=data, headers=headers, auth=auth,
+        response = req_method(url, data=data, params=filters, headers=headers, auth=auth,
                               timeout=timeout, verify=False, stream=False)
 
         return parse_response(response, debug=debug)
@@ -167,13 +172,15 @@ def build_headers(extra_headers=None):
     return headers
 
 
-def build_body_and_url(url, method, data=None, resource_id=None):
-    data = json.dumps(data) if data else None
-
+def build_url(url, method, action=None, resource_id=None, action_id=None):
     if resource_id:
         url += '/%d' % resource_id
+    if action:
+        url += '/%s' % action
+        if action_id:
+            url += '/%d' % action_id
 
-    return data, url
+    return url
 
 
 def parse_response(response, debug=False):
@@ -188,29 +195,29 @@ def parse_response(response, debug=False):
         logging.debug('RESP_HEADERS: %s' % response.headers)
         logging.debug('RESP_CODE: %s' % response.status_code)
 
-    if response.status_code > 300:
-        errors = data
-        exc_data = dict(message=errors.get('ErrorMessage', None),
-                        info=errors.get('ErrorInfo', None),
-                        response_parsed=data,
-                        response=response,
-                        request=response.request,
-                        )
-
-    if response.status_code == 401:
-        raise AuthorizationError(exc_data)
-    elif response.status_code == 403:
-        raise ActionDeniedError(exc_data)
-    elif response.status_code == 404:
-        raise DoesNotExistError(exc_data)
-    elif response.status_code == 422:
-        raise ValidationError(exc_data)
-    elif response.status_code >= 500:
-        logging.error('Critical API error', exc_info=True, extra=exc_data)
-        raise CriticalApiError(errors)
-    elif response.status_code >= 400:
-        logging.error('API error', exc_info=True, extra=exc_data)
-        raise ApiError(exc_data)
+    # if response.status_code > 300:
+    #     errors = data
+    #     exc_data = dict(message=errors.get('ErrorMessage', None),
+    #                     info=errors.get('ErrorInfo', None),
+    #                     response_parsed=data,
+    #                     response=response,
+    #                     request=response.request,
+    #                     )
+    #
+    # if response.status_code == 401:
+    #     raise AuthorizationError(exc_data)
+    # elif response.status_code == 403:
+    #     raise ActionDeniedError(exc_data)
+    # elif response.status_code == 404:
+    #     raise DoesNotExistError(exc_data)
+    # elif response.status_code == 422:
+    #     raise ValidationError(exc_data)
+    # elif response.status_code >= 500:
+    #     logging.error('Critical API error', exc_info=True, extra=exc_data)
+    #     raise CriticalApiError(errors)
+    # elif response.status_code >= 400:
+    #     logging.error('API error', exc_info=True, extra=exc_data)
+    #     raise ApiError(exc_data)
 
     return data
