--- mailjetv3/.gitignore
@@ -1,30 +1,0 @@
-*~
-\#*\#
-/.emacs.desktop
-/.emacs.desktop.lock
-*.elc
-auto-save-list
-tramp
-.\#*
-
-.projectile
-.ropeproject
-.env
-.overcommit.yml
-
-__pycache__/
-*.py[cod]
-*.so
-*.egg-info/
-.installed.cfg
-*.egg
-
-pip-log.txt
-pip-delete-this-directory.txt
-
-.tox/
-.coverage
-.cache
-nosetests.xml
-coverage.xml
-junit*
