--- mailjetv3/MANIFEST.in
@@ -1,3 +1,0 @@
-include *.ini
-include README.md
-include tests/*
