--- mailjetv3/mailjet/__init__.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python
-# coding=utf-8
-
-from client import Client
-
-__version__ = '0.1.0'
-__all__ = ('Client',)
