--- mailjetv3/pytest.ini
@@ -1,3 +1,0 @@
-[pytest]
-norecursedirs=.hg .idea .ropeproject conf .env .git .tox
-python_files=test.py tests.py tests/__init__.py tests/test_*.py
