--- mailjetv3/tox.ini
@@ -1,15 +1,0 @@
-[tox]
-envlist = {py26,py27}
-usedevelop = True
-
-[tox:jenkins]
-envlist = {py26,py27}
-usedevelop = False
-
-[testenv]
-deps =
-    pytest
-    pytest-cov
-
-commands =
-    py.test tests --junitxml=junit-{envname}.xml --cov-report xml --cov mailjet {posargs}
